LayoutTests/http/tests/security/contentSecurityPolicy/1.1/referrer-invalid-https-https-expected.txt

CONSOLE MESSAGE: The 'referrer' Content Security Policy directive has the invalid value "invalid". Valid values are "no-referrer", "no-referrer-when-downgrade", "origin", "origin-when-cross-origin", and "unsafe-url".
Test that invalid is ignored from HTTPS to HTTPS.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ReferrerTest.referrerResult is "https://127.0.0.1:8443/security/contentSecurityPolicy/resources/referrer-test.php?policy=invalid&to=https"
PASS successfullyParsed is true

TEST COMPLETE
